requirements.txt
python-telegram-bot
crewai
langchain
faiss-cpu
python-dotenv
openai
